Se agregaron varias clases de dominio, servicios y controladores, pero el proyecto quedara en stand by de momento

+++ src/main/java/com/arg/ecommerce/service/ProductServiceImpl.java
@@ -1,7 +1,12 @@
 package com.arg.ecommerce.service;
 
 import com.arg.ecommerce.dao.ProductDao;
+import com.arg.ecommerce.dao.ProductSizeDao;
+import com.arg.ecommerce.dao.SizeDao;
 import com.arg.ecommerce.domain.Product;
+import com.arg.ecommerce.domain.ProductSize;
+import com.arg.ecommerce.domain.Size;
+import com.arg.ecommerce.dto.ProductDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -15,8 +20,13 @@ public class ProductServiceImpl implements ProductService{
     @Autowired
     private ProductDao productDao;
 
+    @Autowired
+    private SizeDao sizeDao;
 
-    @Transactional(readOnly = true)
+    @Autowired
+    private ProductSizeDao productSizeDao;
+
+
     @Override
     public List<Product> listProducts() {
         return productDao.findAll();
@@ -30,13 +40,39 @@ public class ProductServiceImpl implements ProductService{
 
     @Transactional
     @Override
-    public void delete(Product product) {
-        productDao.delete(product);
+    public void delete(Long id) {
+        Product product = this.findProductById(id);
+        if(product != null){
+            productDao.delete(product);
+        }
     }
 
     @Transactional(readOnly = true)
     @Override
-    public Product findProductById(Product product) {
-        return productDao.findById(product.getIdProduct()).orElse(null);
+    public Product findProductById(Long id) {
+        return productDao.findById(id).orElse(null);
     }
+
+    @Override
+    public void addSize(Product product, String value) {
+        Size size = sizeDao.findSizeByValue(value);
+        if(size != null){
+            this.createProductSize(product, size);
+        }
+        else{
+            size = new Size();
+            size.setValue(value);
+            sizeDao.save(size);
+            this.createProductSize(product, size);
+        }
+    }
+
+    private void createProductSize(Product product, Size size){
+        ProductSize productSize = new ProductSize();
+        productSize.setProduct(product);
+        productSize.setSize(size);
+        productSizeDao.save(productSize);
+    }
+
+
 }
